Extract generic FileUploader component from KB upload code
