+++ css/includes/_includes.scss
@@ -51,6 +51,7 @@ $is-dark: false !default;
 @import "components/form/form-destination";
 @import "components/form/item-translations";
 @import "components/form/helpdesk-home-config-for-empty-entity";
+@import "components/file-uploader";
 @import "components/fuzzy";
 @import "components/global-menu";
 @import "components/illustration-picker";
