+++ templates/pages/tools/kb/article.html.twig
@@ -283,19 +283,21 @@
                                                 <input type="hidden" name="itemtype" value="KnowbaseItem">
                                                 <input type="hidden" name="items_id" value="{{ item_id }}">
 
-                                                {# Drop Zone #}
-                                                <label class="kb-dropzone" for="kb-document-input">
+                                                {# File uploader component #}
+                                                <label class="file-uploader-dropzone" for="kb-document-input"
+                                                       data-glpi-file-uploader-dropzone>
                                                     <input type="file"
                                                            id="kb-document-input"
                                                            name="files[]"
                                                            multiple
-                                                           class="d-none">
-                                                    <div class="kb-dropzone-content">
-                                                        <i class="ti ti-cloud-upload kb-dropzone-icon"></i>
-                                                        <p class="kb-dropzone-title">
+                                                           class="d-none"
+                                                           data-glpi-file-uploader-input>
+                                                    <div class="file-uploader-dropzone-content">
+                                                        <i class="ti ti-cloud-upload file-uploader-dropzone-icon"></i>
+                                                        <p class="file-uploader-dropzone-title">
                                                             {{ __('Drop files here or click to browse') }}
                                                         </p>
-                                                        <p class="kb-dropzone-hint text-muted">
+                                                        <p class="file-uploader-dropzone-hint text-muted">
                                                             {{ __('PDF, Word, Excel, images and archives are supported') }}
                                                         </p>
                                                         <span class="btn btn-outline-primary btn-sm mt-2">
@@ -305,12 +307,13 @@
                                                     </div>
                                                 </label>
 
-                                                {# File Preview List #}
-                                                <div id="kb-file-preview" class="kb-file-preview mt-3 d-none">
-                                                    <div class="kb-file-list" role="list"></div>
+                                                <div class="file-uploader-preview mt-3 d-none"
+                                                     data-glpi-file-uploader-preview>
+                                                    <div class="file-uploader-list" role="list"
+                                                         data-glpi-file-uploader-list></div>
                                                 </div>
 
-                                                {# Description #}
+                                                {# KB-specific: Description #}
                                                 <div class="mb-3 mt-3">
                                                     <label for="kb-document-description" class="form-label">
                                                         {{ __('Description') }}
@@ -328,7 +331,8 @@
                                                     <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">
                                                         {{ __('Cancel') }}
                                                     </button>
-                                                    <button type="submit" class="btn btn-primary" id="kb-upload-btn" disabled>
+                                                    <button type="submit" class="btn btn-primary"
+                                                            data-glpi-kb-upload-submit disabled>
                                                         <i class="ti ti-upload me-1"></i>
                                                         {{ __('Upload Documents') }}
                                                     </button>
@@ -421,7 +425,7 @@
                     const uploadModule = await import(
                         "{{ js_path('js/modules/Document/UploadController.js') }}"
                     );
-                    new uploadModule.GlpiDocumentUploadController(uploadPane, uploadModal);
+                    new uploadModule.DocumentUploadController(uploadPane, uploadModal);
                     uploadPane.dataset.initialized = 'true';
                 }
             }, { once: true });
